694-i18n - Refactor config. Prepare for something more then just DB
