+++ modules/bots/m0sconiBot/src/main/resources/application.conf
@@ -1,10 +1,12 @@
-mosDB {
+mos {
+  db = {
 
-  driver = "org.sqlite.JDBC"
+    driver = "org.sqlite.JDBC"
 
-  db-name = "../botDB.sqlite3"
-  db-name = ${?DB_NAME}
+    db-name = "../botDB.sqlite3"
+    db-name = ${?DB_NAME}
 
-  url = "jdbc:sqlite:"${mosDB.db-name}
-  url = ${?DB_CONNECTION_URL}
+    url = "jdbc:sqlite:"${mos.db.db-name}
+    url = ${?DB_CONNECTION_URL}
+  }
 }
