+++ modules/bots/richardPHJBensonBot/src/main/resources/application.conf
@@ -1,10 +1,12 @@
-rphjbDB {
+rphjb {
+  db = {
 
-  driver = "org.sqlite.JDBC"
+    driver = "org.sqlite.JDBC"
 
-  db-name = "../botDB.sqlite3"
-  db-name = ${?DB_NAME}
+    db-name = "../botDB.sqlite3"
+    db-name = ${?DB_NAME}
 
-  url = "jdbc:sqlite:"${rphjbDB.db-name}
-  url = ${?DB_CONNECTION_URL}
+    url = "jdbc:sqlite:"${rphjb.db.db-name}
+    url = ${?DB_CONNECTION_URL}
+  }
 }
