+++ modules/bots/xahLeeBot/src/main/resources/application.conf
@@ -1,10 +1,13 @@
-xahDB {
+xah {
 
-  driver = "org.sqlite.JDBC"
+  db = {
 
-  db-name = "../botDB.sqlite3"
-  db-name = ${?DB_NAME}
+    driver = "org.sqlite.JDBC"
 
-  url = "jdbc:sqlite:"${xahDB.db-name}
-  url = ${?DB_CONNECTION_URL}
+    db-name = "../botDB.sqlite3"
+    db-name = ${?DB_NAME}
+
+    url = "jdbc:sqlite:"${xah.db.db-name}
+    url = ${?DB_CONNECTION_URL}
+  }
 }
