+++ modules/bots/youTuboAncheI0Bot/src/main/resources/application.conf
@@ -1,10 +1,13 @@
-ytaiDB {
+ytai {
 
-  driver = "org.sqlite.JDBC"
+  db =  {
 
-  db-name = "../botDB.sqlite3"
-  db-name = ${?DB_NAME}
+    driver = "org.sqlite.JDBC"
 
-  url = "jdbc:sqlite:"${ytaiDB.db-name}
-  url = ${?DB_CONNECTION_URL}
+    db-name = "../botDB.sqlite3"
+    db-name = ${?DB_NAME}
+
+    url = "jdbc:sqlite:"${ytai.db.db-name}
+    url = ${?DB_CONNECTION_URL}
+  }
 }
