+++ modules/main/src/main/resources/application.conf
@@ -24,71 +24,83 @@ main {
     url = ${?MAIN_DB_CONNECTION_URL}
   }
 
-  rphjbDB {
-    driver = "org.sqlite.JDBC"
+  rphjb {
+    db = {
+      driver = "org.sqlite.JDBC"
 
-    # going back to the root of the project from main/target/scala-2.13/
-    # be sure to put here the right path to the db
-    db-name = "../botDB.sqlite3"
-    db-name = ${?RPHJBDB_DB_NAME}
+      # going back to the root of the project from main/target/scala-2.13/
+      # be sure to put here the right path to the db
+      db-name = "../botDB.sqlite3"
+      db-name = ${?RPHJB_DB_NAME}
 
-    url = "jdbc:sqlite:"${main.rphjbDB.db-name}
-    url = ${?RPHJB_DB_CONNECTION_URL}
+      url = "jdbc:sqlite:"${main.rphjb.db.db-name}
+      url = ${?RPHJB_DB_CONNECTION_URL}
+    }
   }
 
-  xahDB {
+  xah {
+    db = {
 
-    driver = "org.sqlite.JDBC"
+      driver = "org.sqlite.JDBC"
 
-    db-name = "../botDB.sqlite3"
-    db-name = ${?XAHDB_DB_NAME}
+      db-name = "../botDB.sqlite3"
+      db-name = ${?XAH_DB_NAME}
 
-    url = "jdbc:sqlite:"${main.xahDB.db-name}
-    url = ${?XAHDB_DB_CONNECTION_URL}
+      url = "jdbc:sqlite:"${main.xah.db.db-name}
+      url = ${?XAH_DB_CONNECTION_URL}
+    }
   }
 
-  ytaiDB {
+  ytai {
+    db = {
 
-    driver = "org.sqlite.JDBC"
+      driver = "org.sqlite.JDBC"
 
-    db-name = "../botDB.sqlite3"
-    db-name = ${?YTAIDB_DB_NAME}
+      db-name = "../botDB.sqlite3"
+      db-name = ${?YTAI_DB_NAME}
 
-    url = "jdbc:sqlite:"${main.ytaiDB.db-name}
-    url = ${?YTAIDB_DB_CONNECTION_URL}
+      url = "jdbc:sqlite:"${main.ytai.db.db-name}
+      url = ${?YTAI_DB_CONNECTION_URL}
+    }
   }
 
-  calaDB {
+  cala {
+    db = {
 
-    driver = "org.sqlite.JDBC"
+      driver = "org.sqlite.JDBC"
 
-    db-name = "../botDB.sqlite3"
-    db-name = ${?CALADB_DB_NAME}
+      db-name = "../botDB.sqlite3"
+      db-name = ${?CALA_DB_NAME}
 
-    url = "jdbc:sqlite:"${main.calaDB.db-name}
-    url = ${?CALADB_DB_CONNECTION_URL}
+      url = "jdbc:sqlite:"${main.cala.db.db-name}
+      url = ${?CALA_DB_CONNECTION_URL}
+    }
   }
 
-  abarDB {
+  abar {
+    db = {
 
-    driver = "org.sqlite.JDBC"
+      driver = "org.sqlite.JDBC"
 
-    db-name = "../botDB.sqlite3"
-    db-name = ${?ABARDB_DB_NAME}
+      db-name = "../botDB.sqlite3"
+      db-name = ${?ABAR_DB_NAME}
 
-    url = "jdbc:sqlite:"${main.abarDB.db-name}
-    url = ${?ABAR_DB_CONNECTION_URL}
+      url = "jdbc:sqlite:"${main.abar.db.db-name}
+      url = ${?ABAR_DB_CONNECTION_URL}
+    }
   }
 
-  mosDB {
+  mos {
+    db = {
 
-    driver = "org.sqlite.JDBC"
+      driver = "org.sqlite.JDBC"
 
-    db-name = "../botDB.sqlite3"
-    db-name = ${?MOSDB_DB_NAME}
+      db-name = "../botDB.sqlite3"
+      db-name = ${?MOS_DB_NAME}
 
-    url = "jdbc:sqlite:"${main.mosDB.db-name}
-    url = ${?MOSDB_DB_CONNECTION_URL}
+      url = "jdbc:sqlite:"${main.mos.db.db-name}
+      url = ${?MOS_DB_CONNECTION_URL}
+    }
   }
 
 }
