+++ modules/main/src/main/scala/com/benkio/main/Config.scala
@@ -1,7 +1,7 @@
 package com.benkio.main
 
 import cats.effect.Async
-import com.benkio.telegrambotinfrastructure.Config as DBConfig
+import com.benkio.telegrambotinfrastructure.initialization.DBConfig
 import pureconfig.*
 
 final case class Config(
