+++ modules/main/src/test/scala/com/benkio/main/ConfigSpec.scala
@@ -7,11 +7,14 @@ class ConfigSpec extends CatsEffectSuite {
 
   test("Config.loadConfig should load the correct configuration with default values") {
     Config.loadConfig[IO].map { config =>
-      assert(config.hostUrl == "0.0.0.0")
-      assert(config.port == 8443)
-      assert(config.webhookCertificate == Some("sbotsCertificatePub.pem"))
-      assert(config.keystorePath == Some("sbotsKeystore.jks"))
-      assert(config.keystorePassword == Some("sbotsKeystorePassword"))
+      assertEquals(config.hostUrl, "0.0.0.0")
+      assertEquals(config.port, 8443)
+      assertEquals(config.webhookCertificate, Some("sbotsCertificatePub.pem"))
+      assertEquals(config.keystorePath, Some("sbotsKeystore.jks"))
+      assertEquals(config.keystorePassword, Some("sbotsKeystorePassword"))
+      assertEquals(config.mainDB.driver, "org.sqlite.JDBC")
+      assertEquals(config.mainDB.dbName, "../botDB.sqlite3")
+      assertEquals(config.mainDB.url, "jdbc:sqlite:../botDB.sqlite3")
     }
   }
 }
