+++ modules/telegramBotInfrastructure/src/test/resources/application.conf
@@ -1,13 +1,14 @@
-main{
+main {
   testInfraDBConf {
-    driver = "org.sqlite.JDBC"
+    db = {
+      driver = "org.sqlite.JDBC"
 
-    # going back to the root of the project from main/target/scala-2.13/
-    # be sure to put here the right path to the db
-    db-name = "testDB.sqlite3"
-    db-name = ${?MAIN_DB_NAME}
+      # going back to the root of the project from main/target/scala-2.13/
+      # be sure to put here the right path to the db
+      db-name = "testDB.sqlite3"
+      db-name = ${?MAIN_DB_NAME}
 
-    url = "jdbc:sqlite:testDB.sqlite3"
-    url = ${?MAIN_DB_CONNECTION_URL}
+      url = "jdbc:sqlite:testDB.sqlite3"
+      url = ${?MAIN_DB_CONNECTION_URL}}
   }
 }
