+++ modules/telegramBotInfrastructure/src/test/scala/com/benkio/telegrambotinfrastructure/ConfigSpec.scala
@@ -1,6 +1,7 @@
 package com.benkio.telegrambotinfrastructure
 
 import cats.effect.IO
+import com.benkio.telegrambotinfrastructure.initialization.Config
 import log.effect.fs2.SyncLogWriter.consoleLogUpToLevel
 import log.effect.LogLevels
 import log.effect.LogWriter
@@ -12,9 +13,9 @@ class ConfigSpec extends CatsEffectSuite {
 
   test("Config.loadConfig should load the correct configuration with default values") {
     Config.loadConfig[IO]("testInfraDBConf").map { config =>
-      assert(config.driver == "org.sqlite.JDBC")
-      assert(config.dbName == "testDB.sqlite3")
-      assert(config.url == "jdbc:sqlite:testDB.sqlite3")
+      assert(config.db.driver == "org.sqlite.JDBC")
+      assert(config.db.dbName == "testDB.sqlite3")
+      assert(config.db.url == "jdbc:sqlite:testDB.sqlite3")
     }
   }
 }
